Remove unused envvars from GitHub Actions
Docker containers and images only get environment variables explicitly
defined in the Dockerfile or `docker` command. Most of the `env` blocks
in our step definitions are therefore noops and can be safely removed.

Similarly, `--env` arguments that are already defined in the Dockerfile
are redundant and also noops. Clean those up too.